Fix spelling of NotificationService
