+++ Avalonia-Ex1-IntroNav/Services/INotificationService.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Avalonia.Controls;
+
+namespace SampleMvvmApp.Services;
+
+public interface INotificationService
+{
+  int NotificationTimeout { get; set; }
+
+  void SetHostWindow(TopLevel window);
+
+  void Show(string title, string message, Action? onClick = null);
+}
