--- Avalonia-Ex1-IntroNav/Services/INotifictionService.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Avalonia.Controls;
-
-namespace SampleMvvmApp.Services
-{
-    public interface INotificationService
-    {
-        int NotificationTimeout { get; set; }
-
-        void SetHostWindow(TopLevel window);
-
-        void Show(string title, string message, Action? onClick = null);
-    }
-}
