+++ Avalonia-Ex3-SplitView/Services/INotificationService.cs
