+++ Avalonia-Ex3-SplitView/Services/NotificationService.cs
